chore: move remaining reference-data files into features/referenceData/
Complete the reference data extraction (de-sloppification C1):
- Move InlineDialog.tsx, ReferenceDataInitialLoadingState.tsx,
  and useReferenceDataManagement.ts from features/classes/ into
  features/referenceData/
- Update all cross-feature imports to direct sibling imports

Lint: clean. Tests: 91/91 pass.
