+++ src/frontend/src/features/referenceData/InlineDialog.tsx
